+++ common-hooks/tls-certificate/internal_tls.go
@@ -3,13 +3,14 @@ package tlscertificate
 import (
 	"context"
 	"fmt"
+	"strings"
 	"time"
 
 	"github.com/deckhouse/module-sdk/pkg"
 	"github.com/deckhouse/module-sdk/pkg/certificate"
 	objectpatch "github.com/deckhouse/module-sdk/pkg/object-patch"
 	"github.com/deckhouse/module-sdk/pkg/registry"
-	tlscertificate "github.com/deckhouse/module-sdk/pkg/tls-certificate"
+	certificatesv1 "k8s.io/api/certificates/v1"
 	"k8s.io/utils/net"
 )
 
@@ -21,6 +22,7 @@ const (
 	// certificate encryption algorithm
 	keyAlgorithm = "ecdsa"
 	keySize      = 256
+	SnapshotKey  = "secret"
 )
 
 var JQFilterTLS = `{
@@ -29,7 +31,8 @@ var JQFilterTLS = `{
     "ca": .data."ca.crt"
 }`
 
-// RegisterInternalTLSHook
+// RegisterInternalTLSHookEM must be used for external modules
+//
 // Register hook which save tls cert in values from secret.
 // If secret is not created hook generate CA with long expired time
 // and generate tls cert for passed domains signed with generated CA.
@@ -39,12 +42,12 @@ var JQFilterTLS = `{
 // Therese tls cert often use for in cluster https communication
 // with service which order tls
 // Clients need to use CA cert for verify connection
-func RegisterInternalTLSHook(conf tlscertificate.GenSelfSignedTLSHookConf) bool {
+func RegisterInternalTLSHookEM(conf GenSelfSignedTLSHookConf) bool {
 	return registry.RegisterFunc(&pkg.HookConfig{
 		OnBeforeHelm: &pkg.OrderedConfig{Order: 5},
 		Kubernetes: []pkg.KubernetesConfig{
 			{
-				Name:       tlscertificate.SnapshotKey,
+				Name:       SnapshotKey,
 				APIVersion: "v1",
 				Kind:       "Secret",
 				NamespaceSelector: &pkg.NamespaceSelector{
@@ -58,10 +61,16 @@ func RegisterInternalTLSHook(conf tlscertificate.GenSelfSignedTLSHookConf) bool 
 				JqFilter: JQFilterTLS,
 			},
 		},
+		Schedule: []pkg.ScheduleConfig{
+			{
+				Name:    "internalTLSSchedule",
+				Crontab: "42 4 * * *",
+			},
+		},
 	}, genSelfSignedTLS(conf))
 }
 
-func genSelfSignedTLS(conf tlscertificate.GenSelfSignedTLSHookConf) func(ctx context.Context, input *pkg.HookInput) error {
+func genSelfSignedTLS(conf GenSelfSignedTLSHookConf) func(ctx context.Context, input *pkg.HookInput) error {
 	var usages []string
 	if conf.Usages == nil {
 		usages = []string{
@@ -88,7 +97,7 @@ func genSelfSignedTLS(conf tlscertificate.GenSelfSignedTLSHookConf) func(ctx con
 
 		cn, sans := conf.CN, conf.SANs(input)
 
-		certs, err := objectpatch.UnmarshalToStruct[certificate.Certificate](input.Snapshots, tlscertificate.SnapshotKey)
+		certs, err := objectpatch.UnmarshalToStruct[certificate.Certificate](input.Snapshots, SnapshotKey)
 		if err != nil {
 			return fmt.Errorf("unmarshal to struct: %w", err)
 		}
@@ -223,3 +232,77 @@ func isOutdatedCA(ca string) (bool, error) {
 
 	return false, nil
 }
+
+// DefaultSANs helper to generate list of sans for certificate
+// you can also use helpers:
+//
+//	ClusterDomainSAN(value) to generate sans with respect of cluster domain (e.g.: "app.default.svc" with "cluster.local" value will give: app.default.svc.cluster.local
+//	PublicDomainSAN(value)
+func DefaultSANs(sans []string) SANsGenerator {
+	return func(input *pkg.HookInput) []string {
+		res := make([]string, 0, len(sans))
+
+		clusterDomain := input.Values.Get("global.discovery.clusterDomain").String()
+		publicDomain := input.Values.Get("global.modules.publicDomainTemplate").String()
+
+		for _, san := range sans {
+			switch {
+			case strings.HasPrefix(san, publicDomainPrefix) && publicDomain != "":
+				san = getPublicDomainSAN(san, publicDomain)
+
+			case strings.HasPrefix(san, clusterDomainPrefix) && clusterDomain != "":
+				san = getClusterDomainSAN(san, clusterDomain)
+			}
+
+			res = append(res, san)
+		}
+		return res
+	}
+}
+
+type GenSelfSignedTLSHookConf struct {
+	// SANs function which returns list of domain to include into cert. Use DefaultSANs helper
+	SANs SANsGenerator
+
+	// CN - Certificate common Name
+	// often it is module name
+	CN string
+
+	// Namespace - namespace for TLS secret
+	Namespace string
+	// TLSSecretName - TLS secret name
+	// secret must be TLS secret type https://kubernetes.io/docs/concepts/configuration/secret/#tls-secrets
+	// CA certificate MUST set to ca.crt key
+	TLSSecretName string
+
+	// Usages specifies valid usage contexts for keys.
+	// See: https://tools.ietf.org/html/rfc5280#section-4.2.1.3
+	//      https://tools.ietf.org/html/rfc5280#section-4.2.1.12
+	Usages []certificatesv1.KeyUsage
+
+	// FullValuesPathPrefix - prefix full path to store CA certificate TLS private key and cert
+	// full paths will be
+	//   FullValuesPathPrefix + .ca  - CA certificate
+	//   FullValuesPathPrefix + .crt - TLS private key
+	//   FullValuesPathPrefix + .key - TLS certificate
+	// Example: FullValuesPathPrefix =  'prometheusMetricsAdapter.internal.adapter'
+	// Values to store:
+	// prometheusMetricsAdapter.internal.adapter.ca
+	// prometheusMetricsAdapter.internal.adapter.crt
+	// prometheusMetricsAdapter.internal.adapter.key
+	// Data in values store as plain text
+	// In helm templates you need use `b64enc` function to encode
+	FullValuesPathPrefix string
+
+	// BeforeHookCheck runs check function before hook execution. Function should return boolean 'continue' value
+	// if return value is false - hook will stop its execution
+	// if return value is true - hook will continue
+	BeforeHookCheck func(input *pkg.HookInput) bool
+}
+
+func (gss GenSelfSignedTLSHookConf) Path() string {
+	return strings.TrimSuffix(gss.FullValuesPathPrefix, ".")
+}
+
+// SANsGenerator function for generating sans
+type SANsGenerator func(input *pkg.HookInput) []string
