+++ common-hooks/tls-certificate/order_certificate.go
@@ -25,12 +25,18 @@ import (
 
 	"github.com/cloudflare/cfssl/helpers"
 	certificatesv1 "k8s.io/api/certificates/v1"
+	v1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	"github.com/deckhouse/module-sdk/pkg"
+	"github.com/deckhouse/module-sdk/pkg/certificate"
 	objectpatch "github.com/deckhouse/module-sdk/pkg/object-patch"
 	"github.com/deckhouse/module-sdk/pkg/registry"
-	tlscertificate "github.com/deckhouse/module-sdk/pkg/tls-certificate"
 )
+
+// CertificateWaitTimeoutDefault controls default amount of time we wait for certificate
+// approval in one iteration.
+const CertificateWaitTimeoutDefault = 1 * time.Minute
 
 type OrderCertificateRequest struct {
 	Namespace  string
@@ -71,7 +77,8 @@ var JQFilterApplyCertificateSecret = `{
     "cert": (if (.data."tls.crt" != null and .data."tls.crt" != "") then .data."tls.crt" else (if (.data."client.crt" != null and .data."client.crt" != "") then .data."client.crt" else null end) end)
 }`
 
-func RegisterOrderCertificateHook(requests []tlscertificate.OrderCertificateRequest) bool {
+// RegisterOrderCertificateHookEM must be used for external modules
+func RegisterOrderCertificateHookEM(requests []OrderCertificateRequest) bool {
 	var namespaces []string
 	var secretNames []string
 	for _, request := range requests {
@@ -101,13 +108,13 @@ func RegisterOrderCertificateHook(requests []tlscertificate.OrderCertificateRequ
 	}, certificateHandler(requests))
 }
 
-func certificateHandler(requests []tlscertificate.OrderCertificateRequest) func(ctx context.Context, input *pkg.HookInput) error {
+func certificateHandler(requests []OrderCertificateRequest) func(ctx context.Context, input *pkg.HookInput) error {
 	return func(ctx context.Context, input *pkg.HookInput) error {
 		return certificateHandlerWithRequests(ctx, input, requests)
 	}
 }
 
-func certificateHandlerWithRequests(ctx context.Context, input *pkg.HookInput, requests []tlscertificate.OrderCertificateRequest) error {
+func certificateHandlerWithRequests(ctx context.Context, input *pkg.HookInput, requests []OrderCertificateRequest) error {
 	publicDomain := input.Values.Get("global.modules.publicDomainTemplate").String()
 	clusterDomain := input.Values.Get("global.discovery.clusterDomain").String()
 
@@ -117,23 +124,23 @@ func certificateHandlerWithRequests(ctx context.Context, input *pkg.HookInput, r
 		// Convert cluster domain and public domain sans
 		for index, san := range request.SANs {
 			switch {
-			case strings.HasPrefix(san, tlscertificate.PublicDomainPrefix) && publicDomain != "":
-				request.SANs[index] = tlscertificate.GetPublicDomainSAN(san, publicDomain)
+			case strings.HasPrefix(san, publicDomainPrefix) && publicDomain != "":
+				request.SANs[index] = getPublicDomainSAN(san, publicDomain)
 
-			case strings.HasPrefix(san, tlscertificate.ClusterDomainPrefix) && clusterDomain != "":
-				request.SANs[index] = tlscertificate.GetClusterDomainSAN(san, clusterDomain)
+			case strings.HasPrefix(san, clusterDomainPrefix) && clusterDomain != "":
+				request.SANs[index] = getClusterDomainSAN(san, clusterDomain)
 			}
 		}
 
 		valueName := fmt.Sprintf("%s.%s", request.ModuleName, request.ValueName)
 
-		certSecrets, err := objectpatch.UnmarshalToStruct[tlscertificate.CertificateSecret](input.Snapshots, "certificateSecrets")
+		certSecrets, err := objectpatch.UnmarshalToStruct[CertificateSecret](input.Snapshots, "certificateSecrets")
 		if err != nil {
 			return fmt.Errorf("unmarshal to struct: %w", err)
 		}
 
 		if len(certSecrets) > 0 {
-			var secret *tlscertificate.CertificateSecret
+			var secret *CertificateSecret
 
 			for _, snapSecret := range certSecrets {
 				if snapSecret.Name == request.SecretName {
@@ -149,14 +156,14 @@ func certificateHandlerWithRequests(ctx context.Context, input *pkg.HookInput, r
 					return err
 				}
 				if !genNew {
-					info := tlscertificate.CertificateInfo{Certificate: string(secret.Cert), Key: string(secret.Key)}
+					info := CertificateInfo{Certificate: string(secret.Cert), Key: string(secret.Key)}
 					input.Values.Set(valueName, info)
 					continue
 				}
 			}
 		}
 
-		info, err := tlscertificate.IssueCertificate(ctx, input, request)
+		info, err := IssueCertificate(ctx, input, request)
 		if err != nil {
 			return err
 		}
@@ -166,7 +173,7 @@ func certificateHandlerWithRequests(ctx context.Context, input *pkg.HookInput, r
 }
 
 // shouldGenerateNewCert checks that the certificate from the cluster matches the order
-func shouldGenerateNewCert(cert string, request tlscertificate.OrderCertificateRequest, durationLeft time.Duration) (bool, error) {
+func shouldGenerateNewCert(cert string, request OrderCertificateRequest, durationLeft time.Duration) (bool, error) {
 	c, err := helpers.ParseCertificatePEM([]byte(cert))
 	if err != nil {
 		return false, fmt.Errorf("certificate cannot parsed: %v", err)
@@ -209,4 +216,102 @@ func arraysAreEqual(a, b []string) bool {
 		}
 	}
 	return true
+}
+
+type CertificateSecret struct {
+	Name string `json:"name"`
+	Key  string `json:"key"`
+	Cert string `json:"cert"`
+}
+
+type CertificateInfo struct {
+	Certificate        string `json:"certificate,omitempty"`
+	Key                string `json:"key,omitempty"`
+	CertificateUpdated bool   `json:"certificate_updated,omitempty"`
+}
+
+func IssueCertificate(ctx context.Context, input *pkg.HookInput, request OrderCertificateRequest) (*CertificateInfo, error) {
+	k8, err := input.DC.GetK8sClient()
+	if err != nil {
+		return nil, fmt.Errorf("can't init Kubernetes client: %v", err)
+	}
+
+	if request.WaitTimeout == 0 {
+		request.WaitTimeout = CertificateWaitTimeoutDefault
+	}
+
+	if len(request.Usages) == 0 {
+		request.Usages = []certificatesv1.KeyUsage{
+			certificatesv1.UsageDigitalSignature,
+			certificatesv1.UsageKeyEncipherment,
+			certificatesv1.UsageClientAuth,
+		}
+	}
+
+	if request.SignerName == "" {
+		request.SignerName = certificatesv1.KubeAPIServerClientSignerName
+	}
+
+	// Delete existing CSR from the cluster.
+	_ = k8.Delete(ctx, &certificatesv1.CertificateSigningRequest{ObjectMeta: metav1.ObjectMeta{Name: request.CommonName}})
+
+	csrPEM, key, err := certificate.GenerateCSR(input.Logger, request.CommonName,
+		certificate.WithGroups(request.Groups...),
+		certificate.WithSANs(request.SANs...))
+	if err != nil {
+		return nil, fmt.Errorf("error generating CSR: %v", err)
+	}
+
+	// Create new CSR in the cluster.
+	csr := &certificatesv1.CertificateSigningRequest{
+		TypeMeta: metav1.TypeMeta{
+			Kind:       "CertificateSigningRequest",
+			APIVersion: "certificates.k8s.io/v1",
+		},
+		ObjectMeta: metav1.ObjectMeta{
+			Name: request.CommonName,
+		},
+		Spec: certificatesv1.CertificateSigningRequestSpec{
+			Request:           csrPEM,
+			Usages:            request.Usages,
+			SignerName:        request.SignerName,
+			ExpirationSeconds: request.ExpirationSeconds,
+		},
+	}
+
+	// Create CSR.
+	err = k8.Create(ctx, csr)
+	if err != nil {
+		return nil, fmt.Errorf("error creating CertificateSigningRequest: %v", err)
+	}
+
+	// Add CSR approved status.
+	csr.Status.Conditions = append(csr.Status.Conditions,
+		certificatesv1.CertificateSigningRequestCondition{
+			Type:    certificatesv1.CertificateApproved,
+			Status:  v1.ConditionTrue,
+			Reason:  "HookApprove",
+			Message: "This CSR was approved by a hook.",
+		})
+
+	// Approve CSR.
+	err = k8.Status().Update(ctx, csr)
+	if err != nil {
+		return nil, fmt.Errorf("error approving of CertificateSigningRequest: %v", err)
+	}
+
+	ctxWTO, cancel := context.WithTimeout(context.Background(), request.WaitTimeout)
+	defer cancel()
+
+	crtPEM, err := certificate.WaitForCertificate(ctxWTO, k8, csr.Name, csr.UID)
+	if err != nil {
+		return nil, fmt.Errorf("%s CertificateSigningRequest was not signed: %v", request.CommonName, err)
+	}
+
+	// Delete CSR.
+	_ = k8.Delete(ctx, &certificatesv1.CertificateSigningRequest{ObjectMeta: metav1.ObjectMeta{Name: request.CommonName}})
+
+	info := CertificateInfo{Certificate: string(crtPEM), Key: string(key), CertificateUpdated: true}
+
+	return &info, nil
 }
