--- pkg/tls-certificate/internal_tls.go
@@ -1,103 +1,0 @@
-/*
-Copyright 2021 Flant JSC
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package tlscertificate
-
-import (
-	"strings"
-
-	certificatesv1 "k8s.io/api/certificates/v1"
-
-	"github.com/deckhouse/module-sdk/pkg"
-)
-
-const (
-	SnapshotKey = "secret"
-)
-
-// DefaultSANs helper to generate list of sans for certificate
-// you can also use helpers:
-//
-//	ClusterDomainSAN(value) to generate sans with respect of cluster domain (e.g.: "app.default.svc" with "cluster.local" value will give: app.default.svc.cluster.local
-//	PublicDomainSAN(value)
-func DefaultSANs(sans []string) SANsGenerator {
-	return func(input *pkg.HookInput) []string {
-		res := make([]string, 0, len(sans))
-
-		clusterDomain := input.Values.Get("global.discovery.clusterDomain").String()
-		publicDomain := input.Values.Get("global.modules.publicDomainTemplate").String()
-
-		for _, san := range sans {
-			switch {
-			case strings.HasPrefix(san, PublicDomainPrefix) && publicDomain != "":
-				san = GetPublicDomainSAN(san, publicDomain)
-
-			case strings.HasPrefix(san, ClusterDomainPrefix) && clusterDomain != "":
-				san = GetClusterDomainSAN(san, clusterDomain)
-			}
-
-			res = append(res, san)
-		}
-		return res
-	}
-}
-
-type GenSelfSignedTLSHookConf struct {
-	// SANs function which returns list of domain to include into cert. Use DefaultSANs helper
-	SANs SANsGenerator
-
-	// CN - Certificate common Name
-	// often it is module name
-	CN string
-
-	// Namespace - namespace for TLS secret
-	Namespace string
-	// TLSSecretName - TLS secret name
-	// secret must be TLS secret type https://kubernetes.io/docs/concepts/configuration/secret/#tls-secrets
-	// CA certificate MUST set to ca.crt key
-	TLSSecretName string
-
-	// Usages specifies valid usage contexts for keys.
-	// See: https://tools.ietf.org/html/rfc5280#section-4.2.1.3
-	//      https://tools.ietf.org/html/rfc5280#section-4.2.1.12
-	Usages []certificatesv1.KeyUsage
-
-	// FullValuesPathPrefix - prefix full path to store CA certificate TLS private key and cert
-	// full paths will be
-	//   FullValuesPathPrefix + .ca  - CA certificate
-	//   FullValuesPathPrefix + .crt - TLS private key
-	//   FullValuesPathPrefix + .key - TLS certificate
-	// Example: FullValuesPathPrefix =  'prometheusMetricsAdapter.internal.adapter'
-	// Values to store:
-	// prometheusMetricsAdapter.internal.adapter.ca
-	// prometheusMetricsAdapter.internal.adapter.crt
-	// prometheusMetricsAdapter.internal.adapter.key
-	// Data in values store as plain text
-	// In helm templates you need use `b64enc` function to encode
-	FullValuesPathPrefix string
-
-	// BeforeHookCheck runs check function before hook execution. Function should return boolean 'continue' value
-	// if return value is false - hook will stop its execution
-	// if return value is true - hook will continue
-	BeforeHookCheck func(input *pkg.HookInput) bool
-}
-
-func (gss GenSelfSignedTLSHookConf) Path() string {
-	return strings.TrimSuffix(gss.FullValuesPathPrefix, ".")
-}
-
-// SANsGenerator function for generating sans
-type SANsGenerator func(input *pkg.HookInput) []string
